docs/jira/autoPM/requirements.txt
langgraph
langchain_openai
beautifulsoup4
requests
python-dotenv